app/requirements.txt
pandas
sqlalchemy
psycopg2-binary
streamlit
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-auth
python-dotenv